Code reorg: Move struct Tbs into identity_assertion.rs
